Move custom attribute retrieval to a reusable Trait.
